# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/assets": {
      "get": {
        "operationId": "AssetsController_getAll",
        "parameters": [
          {
            "name": "location",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Assets"
        ]
      },
      "post": {
        "operationId": "AssetsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Assets"
        ]
      }
    },
...
    "/api/v1/delegations": {
      "get": {
        "operationId": "DelegationsController_getAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Delegations"
        ]
      },
      "post": {
        "operationId": "DelegationsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Delegations"
        ]
      }
    },
    "/api/v1/delegations/{id}/revoke": {
      "patch": {
        "operationId": "DelegationsController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Delegations"
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "CMMS API",
    "description": "The Computerized Maintenance Management System (CMMS) API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateAssetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "description": "The unique ID of the asset"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "The name of the asset"
          },
          "location": {
            "type": "string",
            "maxLength": 100,
            "description": "The location (Product ID) of the asset"
          },
          "department": {
            "type": "string",
            "maxLength": 50,
            "description": "The department of the asset"
          }
        },
        "required": [
          "id",
          "name",
          "location",
          "department"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "employeeId",
          "password"
        ]
      }
    }
  }
}